feat: Scaffold app with a Bun Typescript application (#2)
This application is being written as a Typescript
application powered by the Bun runtime.

This is primarily to make the iteration time _vastly_ faster compared
to, say, a Java based application since build and test times are easier.
Typescript, also, as Anthrophic's team often says, is "on distribution":
a platform and stack that Anthropic models are very good at writing and
reasoning in.

The goal is to speed up, structured development using tools like Claude
Code with explicit guard rails with things like ADRs, CLAUDE.md files
and TDD.